pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dingdong-rag"
version = "0.1.0"
description = "A comprehensive Retrieval-Augmented Generation (RAG) system with advanced document processing, vector storage, reranking, and conversational AI capabilities."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "DingDong RAG Team", email = "contact@dingdong-rag.com"},
]
maintainers = [
    {name = "DingDong RAG Team", email = "contact@dingdong-rag.com"},
]
keywords = ["rag", "retrieval", "augmented", "generation", "document", "processing", "nlp", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11", 
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Linguistic",
]
dependencies = [
    "pymupdf4llm>=0.0.27",
    "pytesseract>=0.3.10",
    "pillow>=10.0.0",
    "pdf2image>=1.17.0",
    "sentence-transformers>=2.2.2",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "tqdm>=4.65.0",
    "openai>=1.0.0",
    "networkx>=3.0",
    "pyvis>=0.3.2",
    "pinecone-client>=3.0.0",
    "rank-bm25>=0.2.2",
    "psutil>=5.9.0",
    "pymupdf>=1.23.0",
    "docling>=2.0.0",
    "google-generativeai>=0.3.0",
    "chromadb>=1.0.15",
    "python-dotenv>=1.0.0",
    "gradio>=4.41.0",
    "cohere>=4.0.0",
    "chonkie>=0.1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "pre-commit>=3.0",
]
docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.0",
    "mkdocstrings[python]>=0.20",
]

[project.scripts]
dingdong-rag = "dingdong_rag.cli.main:main"
dingdong-rag-ui = "dingdong_rag.interactive.app:main"

[project.urls]
"Homepage" = "https://github.com/your-username/dingdong-rag"
"Bug Reports" = "https://github.com/your-username/dingdong-rag/issues"
"Source" = "https://github.com/your-username/dingdong-rag"
"Documentation" = "https://github.com/your-username/dingdong-rag#readme"

[tool.hatch.build.targets.wheel]
packages = ["src/dingdong_rag", "src/dingdong_rag/api", "src/dingdong_rag/caching", "src/dingdong_rag/chat", "src/dingdong_rag/chunking", "src/dingdong_rag/cli", "src/dingdong_rag/core", "src/dingdong_rag/embeddings", "src/dingdong_rag/evaluation", "src/dingdong_rag/interactive", "src/dingdong_rag/metadata", "src/dingdong_rag/monitoring", "src/dingdong_rag/parsing", "src/dingdong_rag/retrieval"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/examples",
    "/README.md",
    "/LICENSE",
    "/pyproject.toml",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src/dingdong_rag --cov-report=term-missing --cov-report=html"

[tool.black]
line-length = 88
target-version = ["py310"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py310" 
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "chromadb.*",
    "pinecone.*",
    "sentence_transformers.*",
    "pytesseract.*",
    "pymupdf4llm.*",
    "rank_bm25.*",
    "docling.*",
    "google.generativeai.*",
    "cohere.*",
    "chonkie.*",
]
ignore_missing_imports = true
